+++ ProjectCAL/calculateVectorMatrix.h
@@ -27,7 +27,7 @@ namespace eigen
     inline std::vector<Eigen::Vector2d> to_vec2(const std::vector<Eigen::Vector3d>& vec3s)
     {
         std::vector<Eigen::Vector2d> res(vec3s.size());
-        for (int i = 0; i < res.size(); ++i)
+        for (int i = 0; i < (int)res.size(); ++i)
             res[i] = to_vec2(vec3s[i]);
         return res;
     }
@@ -48,7 +48,7 @@ namespace eigen
     inline std::vector<Eigen::Vector3d> to_vec3(const std::vector<Eigen::Vector2d>& vec2s)
     {
         std::vector<Eigen::Vector3d> res(vec2s.size());
-        for (int i = 0; i < res.size(); ++i)
+        for (int i = 0; i < (int)res.size(); ++i)
             res[i] = to_vec3(vec2s[i]);
         return res;
     }
